Updated find package implementation to be consistent with control / description yaml parameters
