+++ clearpath_config/links/links.py
@@ -497,8 +497,7 @@ class LinksConfig(BaseConfig):
             mesh: Mesh = None,
             # By Parameters
             name: str = None,
-            visual: str = Mesh.VISUAL,
-            package: str = Mesh.PACKAGE,
+            visual: dict = Mesh.VISUAL,
             parent: str = Accessory.PARENT,
             xyz: List[float] = Accessory.XYZ,
             rpy: List[float] = Accessory.RPY,
@@ -512,7 +511,6 @@ class LinksConfig(BaseConfig):
             mesh = Mesh(
                 name=name,
                 visual=visual,
-                package=package,
                 parent=parent,
                 xyz=xyz,
                 rpy=rpy,
