+++ clearpath_config/platform/extras.py
@@ -25,10 +25,9 @@
 # CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 # ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
-import os
-
 from clearpath_config.common.types.config import BaseConfig
 from clearpath_config.common.types.file import File
+from clearpath_config.common.types.package_path import PackagePath
 from clearpath_config.common.types.platform import Platform
 from clearpath_config.common.utils.dictionary import (
     flatten_dict,
@@ -139,7 +138,6 @@ class ROSParameterDefaults:
 class ExtrasConfig(BaseConfig):
 
     EXTRAS = "extras"
-    URDF_PACKAGE = "urdf_package"
     URDF = "urdf"
     ROS_PARAMETERS = "ros_parameters"
 
@@ -156,7 +154,6 @@ class ExtrasConfig(BaseConfig):
 
     TEMPLATE = {
         EXTRAS: {
-            URDF_PACKAGE: URDF_PACKAGE,
             URDF: URDF,
             ROS_PARAMETERS: {
                 PLATFORM_VELOCITY_CONTROLLER: {
@@ -178,16 +175,14 @@ class ExtrasConfig(BaseConfig):
     KEYS[ROS_PARAMETERS] = ".".join([EXTRAS, ROS_PARAMETERS])
 
     DEFAULTS = {
-        URDF_PACKAGE: "",
-        URDF: "empty.urdf.xacro",
+        URDF: "",
         ROS_PARAMETERS: ROSParameterDefaults(BaseConfig.get_platform_model()),
     }
 
     def __init__(
             self,
             config: dict = {},
-            urdf_package: str = DEFAULTS[URDF_PACKAGE],
-            urdf: str = DEFAULTS[URDF],
+            urdf: dict = DEFAULTS[URDF],
             ros_parameters: dict = {},
             ) -> None:
         # ROS Parameter Setter Template
@@ -204,14 +199,12 @@ class ExtrasConfig(BaseConfig):
         }
         # Setter Template
         self.setters = {
-            self.KEYS[self.URDF_PACKAGE]: ExtrasConfig.urdf_package,
             self.KEYS[self.URDF]: ExtrasConfig.urdf,
             self.KEYS[self.ROS_PARAMETERS]: ExtrasConfig.ros_parameters,
         }
         # Initialization
         self._init_ros_parameter()
         self._config = {}
-        self.urdf_package = urdf_package
         self.urdf = urdf
         self.ros_parameters = ros_parameters
         # Set from Config
@@ -222,39 +215,20 @@ class ExtrasConfig(BaseConfig):
             self._update_ros_parameter()
 
     @property
-    def urdf_package(self) -> str:
-        urdf_package = (None if self._is_default(self._urdf_package, self.URDF_PACKAGE)
-                        else str(self._urdf_package))
-        self.set_config_param(
-            key=self.KEYS[self.URDF_PACKAGE],
-            value=urdf_package
-        )
-        return urdf_package
-
-    @urdf_package.setter
-    def urdf_package(self, value: str) -> None:
-        if value is None or value == "None":
-            return
-        self._urdf_package = value
-
-    @property
-    def urdf(self) -> str:
-        urdf = None if self._is_default(self._urdf, self.URDF) else str(self._urdf)
+    def urdf(self) -> dict:
+        urdf = None if self._is_default(self._urdf, self.URDF) else dict(self._urdf.to_dict())
         self.set_config_param(
             key=self.KEYS[self.URDF],
-            value=urdf
+            value=urdf,
         )
         return urdf
 
     @urdf.setter
-    def urdf(self, value: str) -> None:
+    def urdf(self, value: dict) -> None:
         if value is None or value == "None":
             return
-        elif self._urdf_package:
-            self._urdf = os.path.join("$(find " + self._urdf_package + ")",
-                                      File.clean(path=str(value), make_abs=False))
-        else:
-            self._urdf = File.clean(path=str(value), make_abs=True)
+        self._urdf = PackagePath()
+        self._urdf.from_dict(value)
 
     def _is_default(self, curr: str, key: str) -> bool:
         return curr == str(File(self.DEFAULTS[key]))
